LPS-74032 not dyslexic after all. Matching the exported package com.liferay.portal.search.contributor
