+++ modules/apps/foundation/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/contributor/document/ExpandoBridgeDocumentContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.internal.document.contributor;
+package com.liferay.portal.search.internal.contributor.document;
 
 import com.liferay.expando.kernel.util.ExpandoBridgeIndexer;
 import com.liferay.portal.kernel.model.BaseModel;
